refactor: env 쿠키 기반 콜백 URL 해석 로직을 공용 컴포넌트로 추출
