+++ src/main/java/com/groute/groute_server/auth/service/oauth/OAuth2LoginSuccessHandler.java
@@ -4,13 +4,10 @@ import java.io.IOException;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseCookie;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
@@ -32,9 +29,8 @@ import lombok.extern.slf4j.Slf4j;
  * deliver가 이미 Set-Cookie로 refresh를 전달했으므로 query에는 access만 싣는다. 본문 모드(local)에서는 refresh도 query에 함께
  * 실어 프론트가 즉시 사용할 수 있게 한다.
  *
- * <p>리다이렉트 대상 콜백은 {@link OAuth2EnvAwareAuthorizationRequestResolver}가 OAuth2 시작 시 발행한 {@code
- * oauth_env} 쿠키로 결정한다. 쿠키 값이 {@link AuthProperties#callback()} 화이트리스트에 있으면 그 URL로, 없거나 누락되면 default
- * env로 fallback. 사용 후에는 즉시 만료 처리(Max-Age=0)해 round-trip 흔적을 지운다.
+ * <p>리다이렉트 대상 콜백은 {@link OAuthCallbackUrlResolver}에 위임한다. resolver가 {@code oauth_env} 쿠키로 매칭 URL을
+ * 결정하고 사용 후 쿠키를 만료시킨다.
  *
  * <p>인가 코드 교환 중 잠시 사용된 세션은 더 이상 필요 없으므로 redirect 직전에 invalidate하여 서버 상태를 JWT-only로 복귀시킨다.
  *
@@ -52,6 +48,7 @@ public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
     private final RefreshTokenRepository refreshTokenRepository;
     private final TokenDeliveryService tokenDeliveryService;
     private final AuthProperties authProperties;
+    private final OAuthCallbackUrlResolver oAuthCallbackUrlResolver;
 
     @Override
     public void onAuthenticationSuccess(
@@ -70,7 +67,7 @@ public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
             throw new IllegalStateException("accessToken must not be null after deliver");
         }
 
-        String callbackUrl = resolveCallbackUrl(request, response);
+        String callbackUrl = oAuthCallbackUrlResolver.resolveAndExpire(request, response);
         // 쿠키 모드(prod)면 deliver가 이미 Set-Cookie로 refresh를 전달했으므로 query에는 싣지 않는다.
         String refreshForQuery =
                 authProperties.refreshToken().cookieEnabled() ? null : refreshToken;
@@ -83,45 +80,6 @@ public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
                 principal.getProvider(),
                 callbackUrl);
         response.sendRedirect(redirectUrl);
-    }
-
-    private String resolveCallbackUrl(HttpServletRequest request, HttpServletResponse response) {
-        String env = readEnvCookie(request);
-        String url = (env != null) ? authProperties.callbackUrlFor(env) : null;
-        if (url == null) {
-            if (env != null) {
-                log.warn("oauth_env 쿠키 값이 등록되지 않은 env, default로 fallback: env={}", env);
-            }
-            url = authProperties.defaultCallbackUrl();
-        }
-        expireEnvCookie(request, response);
-        return url;
-    }
-
-    private static String readEnvCookie(HttpServletRequest request) {
-        Cookie[] cookies = request.getCookies();
-        if (cookies == null) {
-            return null;
-        }
-        for (Cookie c : cookies) {
-            if (OAuth2EnvAwareAuthorizationRequestResolver.ENV_COOKIE_NAME.equals(c.getName())) {
-                String value = c.getValue();
-                return (value == null || value.isBlank()) ? null : value;
-            }
-        }
-        return null;
-    }
-
-    private static void expireEnvCookie(HttpServletRequest request, HttpServletResponse response) {
-        ResponseCookie cookie =
-                ResponseCookie.from(OAuth2EnvAwareAuthorizationRequestResolver.ENV_COOKIE_NAME, "")
-                        .maxAge(0)
-                        .httpOnly(true)
-                        .secure(request.isSecure())
-                        .sameSite("Lax")
-                        .path("/")
-                        .build();
-        response.addHeader(HttpHeaders.SET_COOKIE, cookie.toString());
     }
 
     private String buildRedirectUrl(String callbackUrl, String accessToken, String refreshToken) {
